requirements.txt
# Tested with Python 3.9.7
adal==1.2.7
azure-common==1.1.27
azure-core==1.20.1
azure-graphrbac==0.61.1
azure-identity==1.7.1
azure-mgmt-authorization==2.0.0
azure-mgmt-core==1.3.0
azure-mgmt-subscription==1.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.0
fire==0.4.0
idna==3.3
isodate==0.6.0
msal==1.16.0
msal-extensions==0.3.0
msrest==0.6.21
msrestazure==0.6.4
oauthlib==3.1.1
portalocker==1.7.1
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
termcolor==1.1.0
urllib3==1.26.7
